[REFAC]: Move time constants and token expiry definitions to a new constants file for better organization + remove useless command in package.json
